Resolved merge conflicts, moved to foundation package
