+++ feature/tabs/src/main/java/com/epmedu/animeal/tabs/presentation/HomeScreen.kt
@@ -1,7 +1,11 @@
 package com.epmedu.animeal.tabs.presentation
 
 import android.Manifest
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
@@ -15,6 +19,7 @@ import com.epmedu.animeal.base.theme.Typography
 import com.epmedu.animeal.extensions.isPermanentlyDenied
 import com.epmedu.animeal.extensions.launchAppSettings
 import com.epmedu.animeal.foundation.button.AnimealButton
+import com.epmedu.animeal.foundation.switch.AnimealSwitch
 import com.epmedu.animeal.resources.R
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
 import com.google.accompanist.permissions.rememberPermissionState
@@ -44,6 +49,12 @@ internal fun HomeScreen() {
 @Composable
 private fun Maps() {
     Box(modifier = Modifier.fillMaxSize()) {
+        AnimealSwitch(
+            modifier = Modifier
+                .align(alignment = Alignment.TopCenter)
+                .padding(top = 24.dp),
+            onTabSelected = {}
+        )
         Text(text = stringResource(R.string.home), modifier = Modifier.align(Alignment.Center))
     }
 }
@@ -51,7 +62,10 @@ private fun Maps() {
 @Composable
 private fun RationaleText() {
     Box(modifier = Modifier.fillMaxSize()) {
-        Text(text = stringResource(R.string.location_rationale_text), modifier = Modifier.align(Alignment.Center))
+        Text(
+            text = stringResource(R.string.location_rationale_text),
+            modifier = Modifier.align(Alignment.Center)
+        )
     }
 }
